various changes
refactored file structure
placeholder files
import error on line 5 in main.py
