+++ main/cbmcalculator/headings.py
@@ -1,0 +1,19 @@
+"""
+Author: Elvis Obero-Atkins
+Last Edited by: Elvis Obero-Atkins
+"""
+def productdetails_headings():
+    """returns the headings of the product-details sheet
+    Returns:
+        list: stores all the headings for each of the columns
+    """
+    return ['partNo', 'barcode', 'sku', 'productTitle', 'pWidth', 'pHeight', 'pDepth', 'icWidth', 'icHeight',
+            'icDepth', 'icWeight', 'icQty', 'ocWidth', 'ocHeight', 'ocDepth', 'ocWeight', 'ocQty']  # 17 (0-16) Items in list
+
+
+def shippingdetails_headings():
+    """returns the headings of the pallet-shipping sheet
+    Returns:
+        list: stores all the headings for each of the columns
+    """
+    return ['name', 'length', 'width', 'height', 'cbm', 'maxWeight', 'price']  # 7 (0-6) Items in list
