+++ main/cli-app.py
@@ -58,7 +58,7 @@ def main():
     """
 
     # this is unnecessary but cool
-    f = open(".main\welcome.txt", "r")
+    f = open("main/cli-app.txt", "r")
     print(f.read())
     f.close()
