+++ main/odoo-module.py
@@ -1,0 +1,8 @@
+"""
+Author: Elvis Obero-Atkins
+Last Edited by: Elvis Obero-Atkins
+
+This py script is NOT REQUIRED for yandiya-cbm-library to function
+
+It is odoo module written in python designed to interact with the yandiya-cbm-library.
+"""
